만료일 입력 부분을 컴포넌트로 분리
